protection
create listing, delete listing, update listing requires authorisation;

+++ backend/src/controllers/product.js
@@ -46,11 +46,15 @@ createProduct = async (req, res) => {
   const client = await pool.connect();
   try {
     const { productname, description, price, sellerid } = req.body;
-    const data = await client.query(
-      "INSERT INTO products(product_name, description, price, seller_uuid) VALUES ($1, $2, $3, $4)",
-      [productname, description, price, sellerid]
-    );
-    res.json({ status: "success", msg: "product created" });
+    if (req.decoded.uuid === sellerid) {
+      const data = await client.query(
+        "INSERT INTO products(product_name, description, price, seller_uuid) VALUES ($1, $2, $3, $4)",
+        [productname, description, price, sellerid]
+      );
+      res.json({ status: "success", msg: "product created" });
+    } else {
+      throw new Error();
+    }
   } catch (error) {
     console.error(error.message);
     res.status(400).json({ status: "error", msg: "create error" });
@@ -61,11 +65,15 @@ deleteProductById = async (req, res) => {
   const client = await pool.connect();
   try {
     const { productid, sellerid } = req.body;
-    const data = await client.query(
-      "DELETE FROM products WHERE uuid = $1 AND seller_uuid = $2",
-      [productid, sellerid]
-    );
-    res.json({ status: "success", msg: "product deleted" });
+    if (req.decoded.uuid === sellerid) {
+      const data = await client.query(
+        "DELETE FROM products WHERE uuid = $1 AND seller_uuid = $2",
+        [productid, sellerid]
+      );
+      res.json({ status: "success", msg: "product deleted" });
+    } else {
+      throw new Error();
+    }
   } catch (error) {
     console.error(error.message);
     res.status(400).json({ status: "error", msg: "delete error" });
@@ -76,28 +84,33 @@ updateProductById = async (req, res) => {
   const client = await pool.connect();
   try {
     const { uuid, sellerid, productname, description, price } = req.body;
-    const updates = [];
-    const values = [uuid, sellerid];
-    let query = "UPDATE products SET";
+    if (req.decoded.uuid === sellerid) {
+      const updates = [];
+      const values = [uuid, sellerid];
+      let query = "UPDATE products SET";
 
-    if (productname) {
-      values.push(productname);
-      updates.push(` product_name = $${values.length}`);
-    }
-    if (description) {
-      values.push(description);
-      updates.push(` description = $${values.length}`);
-    }
-    if (price) {
-      values.push(price);
-      updates.push(` price = $${values.length}`);
-    }
+      if (productname) {
+        values.push(productname);
+        updates.push(` product_name = $${values.length}`);
+      }
+      if (description) {
+        values.push(description);
+        updates.push(` description = $${values.length}`);
+      }
+      if (price) {
+        values.push(price);
+        updates.push(` price = $${values.length}`);
+      }
 
-    query +=
-      updates.join(", ") + ` WHERE uuid = $1 AND seller_uuid = $2 RETURNING *`;
+      query +=
+        updates.join(", ") +
+        ` WHERE uuid = $1 AND seller_uuid = $2 RETURNING *`;
 
-    const data = await pool.client(query, values);
-    res.json({ status: "success", msg: "product updated" });
+      const data = await pool.client(query, values);
+      res.json({ status: "success", msg: "product updated" });
+    } else {
+      throw new Error();
+    }
   } catch (error) {
     console.error(error.message);
     res.status(400).json({ status: "error", msg: "update error" });
